hushh_mcp/agents/relationship_memory/requirements_api.txt
# FastAPI Backend Requirements for Proactive Relationship Manager

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for external API calls
httpx==0.25.2
requests==2.31.0

# Existing agent dependencies (if not already installed)
langchain-google-genai>=0.1.0
langgraph>=0.0.40
cryptography>=41.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0